chore: fix invalid function location
