+++ src/utils/index.ts
@@ -3,6 +3,9 @@ import deb from '../deb-build';
 import fsLegacy, {promises as fs} from 'fs';
 import rimraf from 'rimraf';
 import copy from 'recursive-copy';
+import { release } from 'os';
+import child_process from 'child_process';
+import tmpjs from 'tmp';
 
 export interface FADeConfiguration {
     name: String,
@@ -88,6 +91,34 @@ export function checkBlacklist(blacklist: Array<String>, name: string): boolean 
 		});
 	}
 	return true;
+}
+export async function openEditor(filename: String, filedata: string): Promise<String> {
+    if(!process.env.EDITOR) {
+        if(process.platform === "win32") {
+            if(release().split('.')[0] == "10" && parseInt(release().split('.')[2]) >= 17763) {
+                console.warn("[FADe] %EDITOR% not set, defaulting to notepad.exe");
+				process.env.EDITOR = "notepad.exe"
+            }else {
+				throw new Error(`[FADe] %EDITOR% not set and Your notepad.exe dosen't support LF Ending.
+Please download your preferred editor from the Internet. We recommend vim or nano
+ - Vim: https://www.vim.org/download.php#pc
+ - Nano: https://www.nano-editor.org/dist/win32-support/
+ 
+Put downloaded binary into C:\\Windows\\system32 or your working directory, and Please type before run FADe:
+> set EDITOR=(binary).exe`);
+			}
+		}else{
+            console.warn("[FADe] $EDITOR not set, defaulting to vi");
+            process.env.EDITOR = "vi"
+        }
+    }
+    let tmpfile = tmpjs.tmpNameSync();
+    console.log(`[FADe] Opening ${filename} with $EDITOR.`);
+    await fs.writeFile(tmpfile, filedata);
+    await child_process.spawn(process.env.EDITOR, [tmpfile], { stdio: 'inherit', detached: true});
+    let return_val = (await fs.readFile(tmpfile)).toString();
+    await fs.unlink(tmpfile);
+    return return_val;
 }
 
 export async function stubCreateDeb(path: string, host: boolean, output?: string): Promise<boolean> {
